Extract StubTranslator to its own class
