+++ src/Test/AbstractWidgetTestCase.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace Sonata\Form\Test;
 
+use Sonata\Form\Fixtures\StubTranslator;
 use Symfony\Bridge\Twig\Extension\FormExtension;
 use Symfony\Bridge\Twig\Extension\TranslationExtension;
 use Symfony\Bridge\Twig\Form\TwigRendererEngine;
@@ -20,30 +21,9 @@ use Symfony\Component\Form\FormRenderer;
 use Symfony\Component\Form\FormView;
 use Symfony\Component\Form\Test\TypeTestCase;
 use Symfony\Component\Security\Csrf\CsrfTokenManagerInterface;
-use Symfony\Component\Translation\TranslatorInterface;
 use Twig\Environment;
 use Twig\Loader\FilesystemLoader;
 use Twig\RuntimeLoader\FactoryRuntimeLoader;
-
-final class StubTranslator implements TranslatorInterface
-{
-    public function trans($id, array $parameters = [], $domain = null, $locale = null)
-    {
-        return '[trans]'.strtr($id, $parameters).'[/trans]';
-    }
-
-    public function transChoice($id, $number, array $parameters = [], $domain = null, $locale = null)
-    {
-    }
-
-    public function setLocale($locale)
-    {
-    }
-
-    public function getLocale()
-    {
-    }
-}
 
 /**
  * Base class for tests checking rendering of form widgets.
